% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bayesian_mte}
\alias{bayesian_mte}
\title{implements the Bayesian marginal treatment effect approach described in Opper (2022).}
\usage{
bayesian_mte(
  outcome,
  treatment_choice,
  treatment_assignment,
  grid_length = 0.01,
  demean = FALSE,
  full_bayes = TRUE,
  input_hypers = NULL,
  log_prior = NULL,
  hyperparameter_draws = 10000,
  extrapolation_uncertainty = FALSE,
  statistical_uncertainty = FALSE
)
}
\arguments{
\item{outcome}{is the outcome; input should be a vector.}

\item{treatment_choice}{is the treatment status; input should be a vector.}

\item{treatment_assignment}{is the treatment assignment; input should be a vector.}

\item{grid_length}{is the grid length of eta; default is 0.01.}

\item{demean}{determines whether to remove the mean from the outcome before the analysis; default is FALSE.}

\item{full_bayes}{determines to integrate over the prior distribution of hyperparameters (if TRUE) or to estimate the hyperparameters (if FALSE); default is true.}

\item{input_hypers}{hyperparameters; default is NULL in which they are estimated.}

\item{log_prior}{is a function that specifies prior -- see documentation for description of function; default is NULL.}

\item{hyperparameter_draws}{is number of draws from the hyper-prior distribution; default is 10,000.}

\item{extrapolation_uncertainty}{determines whether the observed moments are assumed to be estimated without error (if TRUE) or with error (if FALSE);  default is FALSE.}

\item{statistical_uncertainty}{determines whether the reported standard deviations are the standard errors of the MAP (if TRUE) or the posterior standard deviation (if FALSE);  default is FALSE.}
}
\value{
ATE is the posterior mean (ATE$mean) and variance (ATE$variance) of the average treatment effect

LATE, ATATE, NTATE are identical to ATE, but for the local average treatment effect (LATE), always taker average treatment effect (ATATE), and never taker average treatment effect (NTATE)

prior_plot, posterior_plot are plots showing 10 random functions generated according to the estimated hyperparameter. Prior plot does not condition on the observed moments, while posterior plot does. Only returned when full_bayes == FALSE.

MTE, MTE_plot are the posterior mean/variance of the marginal treatment effect function at each point on the eta-grid. MTE is a matrix with the information and MTE_plot returns a graph with the posterior mean and 95\% CI. Only returned when full_bayes == FALSE.

predictions_plot is a plot showing the posterior mean and 95\% CI of the m_1 and m_0 functions, as defined in Opper (2022), along with the observed moments. Only returned when full_bayes == FALSE.

predictions_plot is a plot showing the posterior mean and 95\% CI of the m_1 and m_0 functions, as defined in Opper (2022), along with the observed moments. Only returned when full_bayes == FALSE.

hypers is a list of the hyperparameters that were used. Only returned when full_bayes == FALSE.

posteriors_and_priors is a graph showing the prior and posterior distribution of the hyperparameters. Only returned when full_bayes == TRUE.

number_of_hypers_used is the number of hyperparameters that are "accepted" by the accept/reject algorithm. Only returned when full_bayes == TRUE.
}
\description{
implements the Bayesian marginal treatment effect approach described in Opper (2022).
}
